requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.1
django-autoslug==1.9.9
django-colorfield==0.10.1
Pillow==10.0.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
